Update: theme switcher to a component
